CrewQuery v.01
imdbTest was the first module attempting to search movies for crew details. Runs into KeyValue errors for movies without composers or cinematographers or other attributes. Even though I've excepted the  error, it still doesn't continue past it, so I had to try the module named imdbKeyValueSolve, where I got it to return entire cast and crew lists for any movie you search.

It's not the most efficient code, I've tried the more efficient way at the bottom of the KeyValueSolve file but it runs into index errors for some reason.