CrewCompareV.02
TO DO item near the bottom of the file. Still need to write the compare function. I need to compare the names of the objects in movie1 and movie2list to see if they're the same and just spit out whichever ones match. It is currently running into a TypeError where the 'Person' object isn't callable. And both of the lists are filled with 'Person' objects. SO I may have to edit the big ugly lines of if statements. Or find some other crafty way of extracting the strings of names from the person objects